requirements.txt
requests==2.18.2
Fabric3==1.13.1.post1
